Refactor specification code and Project names. added Domain And Repositories Projects

+++ src/PGP.Infrastrucutre.Framework.Repositories/Entity/IEntity.cs
@@ -11,7 +11,7 @@
         /// <value>
         /// The identifier.
         /// </value>
-        int Id
+        long Id
         {
             get; set;
         }
